+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# 0.1.1
+
+## The Box Model Update
+
+The box model and it's relevant calculations have been completely re-written.
+
+* Margins now stack and work on all instances of Grids ( Grids, Rows & Columns )
+* A new property has been added for content called **[INSTANCE].content.x** & **[INSTANCE].content.y**
+* Padding can now be adjusted, and will modify the **content** properties of an instance
+* Some sizing & positioning issues were addressed in the Box Model Update
+
+
 # 0.1.0
 
 Early commit of GLayout. Contains tools for building HUD's using a grid system.